+++ src/main/java/com/demo/race_condition_demo/dto/TransferRequestDto.java
@@ -1,6 +1,7 @@
 package com.demo.race_condition_demo.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record TransferRequestDto(Double amount, UUID sourceAccountId, UUID targetAccountId) {
+public record TransferRequestDto(BigDecimal amount, UUID sourceAccountId, UUID targetAccountId) {
 }
