+++ src/test/java/com/demo/race_condition_demo/faker/AccountEntityFaker.java
@@ -3,11 +3,12 @@ package com.demo.race_condition_demo.faker;
 import com.demo.race_condition_demo.entities.AccountEntity;
 import com.demo.race_condition_demo.entities.UserEntity;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class AccountEntityFaker {
 
-    public static AccountEntity createSourceAccount(UUID accountId, UserEntity sourceUser, double balance) {
+    public static AccountEntity createSourceAccount(UUID accountId, UserEntity sourceUser, BigDecimal balance) {
         return new AccountEntity(
                 accountId,
                 "03412-9",
@@ -16,7 +17,7 @@ public class AccountEntityFaker {
                 sourceUser);
     }
 
-    public static AccountEntity createTargetAccount(UUID accountId, UserEntity targetUser, double balance) {
+    public static AccountEntity createTargetAccount(UUID accountId, UserEntity targetUser, BigDecimal balance) {
         return new AccountEntity(accountId,
                 "02325-1",
                 "7743",
